Convert XlsConfiguration to AutoConfiguration
Codebase was reconfigured to better utilize Spring Boot's autoconfiguration functionality. The XlsConfiguration class is now marked with the @AutoConfiguration annotation instead of @Configuration. The change applies to all tests using @SpringBootTest, which now specifically import this configuration. This improves integration testing flexibility and can potentially simplify the setup for consumers of this library.